IN-196 fix: move instance to global
