+++ src/common/api/index.ts
@@ -1,1 +1,2 @@
+export * from './instance'
 export * from './locations.api'
