+++ src/common/api/instance.endpoints.ts
@@ -1,0 +1,3 @@
+export const EndpointsToken = {
+  updateToken: '/v1/auth/update-tokens',
+}
