+++ src/common/api/instance.ts
@@ -1,9 +1,14 @@
-import { Paths } from '@/common'
+import { Paths, setToLocalStorage } from '@/common'
 import { StatusCode, StorageKeys } from '@/common/enums'
 import { Environments } from '@/common/enviroments'
 import { clearAllData } from '@/common/utils/clearAllData'
-import { Endpoints, authStore } from '@/features/auth'
-import axios, { AxiosResponse, isAxiosError } from 'axios'
+import axios, {
+  AxiosResponse,
+  InternalAxiosRequestConfig,
+  isAxiosError,
+} from 'axios'
+
+import { EndpointsToken } from './instance.endpoints'
 
 export const instance = axios.create({
   baseURL: Environments.API_URL,
@@ -12,6 +17,26 @@ export const instance = axios.create({
   },
   withCredentials: true,
 })
+
+const updateToken = async (params: InternalAxiosRequestConfig | undefined) => {
+  try {
+    if (localStorage.getItem(StorageKeys.AccessToken)) {
+      const newToken = await instance
+        .post(EndpointsToken.updateToken)
+        .then((res) => res.data.accessToken)
+
+      setToLocalStorage(StorageKeys.AccessToken, newToken)
+
+      if (params) {
+        params.headers.Authorization = `Bearer ${newToken}`
+
+        return await instance.request(params)
+      }
+    }
+  } catch (error) {
+    return Promise.reject(error)
+  }
+}
 
 //TODO фикс появления ошибки, на первом запросе при обновлении токена
 
@@ -22,13 +47,13 @@ instance.interceptors.response.use(
   async (error: unknown) => {
     if (isAxiosError(error)) {
       if (error.response?.status === StatusCode.Unauthorized) {
-        if (error.config?.url === Endpoints.updateToken) {
+        if (error.config?.url === EndpointsToken.updateToken) {
           await clearAllData(Paths.signIn)
 
           return Promise.reject(error)
         } else {
           try {
-            return await authStore.updateToken(error.config)
+            return await updateToken(error.config)
           } catch (updateError) {
             return Promise.reject(updateError)
           }
