+++ src/common/index.ts
@@ -1,4 +1,3 @@
-export * from './api'
 export * from './components'
 export * from './enviroments'
 export * from './form'
