+++ src/features/auth/api/auth.endpoints.ts
@@ -9,5 +9,4 @@ export const Endpoints = {
   login: '/v1/auth/login',
   logout: '/v1/auth/logout',
   me: '/v1/auth/me',
-  updateToken: '/v1/auth/update-tokens',
 }
