+++ src/features/auth/api/index.ts
@@ -1,4 +1,3 @@
 export * from './auth.api'
 export * from './auth.endpoints'
 export * from './authApi.types'
-export * from 'src/features/auth/api/instance'
