+++ src/features/profile/settings/generalInfo/api/profile.api.ts
@@ -1,4 +1,4 @@
-import { instance } from '@/features/auth'
+import { instance } from '@/common/api'
 import { UpdatedProfile, UserProfile } from '@/features/profile'
 import { ProfileEndpoints } from '@/features/profile/settings/generalInfo/api/profile.endpoints'
 import { AxiosInstance, AxiosResponse } from 'axios'
